excellent

+++ extensions/brute_force.cpp
@@ -23,13 +23,13 @@ static inline T* npmem(py::array_t<T> numpy_array)
 
 struct EuclideanDistance
 {
-    double operator()(const double *p, const double *q, int d)
+    float operator()(const float *p, const float *q, int d)
     {
-        double val = 0;
+        float val = 0;
 
         for (int i = 0; i < d; ++i)
         {
-            double delta = p[i] - q[i];
+            float delta = p[i] - q[i];
             val += delta * delta;
         }
 
@@ -39,9 +39,9 @@ struct EuclideanDistance
 
 struct ManhattanDistance
 {
-    double operator()(const double *p, const double *q, int d)
+    float operator()(const float *p, const float *q, int d)
     {
-        double val = 0;
+        float val = 0;
 
         for (int i = 0; i < d; ++i)
         {
@@ -54,9 +54,9 @@ struct ManhattanDistance
 
 struct CosineDistance
 {
-    double operator()(const double *p, const double *q, int d)
+    float operator()(const float *p, const float *q, int d)
     {
-        double val = 0;
+        float val = 0;
 
         for (int i = 0; i < d; ++i)
             val += p[i]*q[i];
@@ -67,9 +67,9 @@ struct CosineDistance
 
 struct ChebyshevDistance
 {
-    double operator()(const double *p, const double *q, int d)
+    float operator()(const float *p, const float *q, int d)
     {
-        double val = 0;
+        float val = 0;
 
         for (int i = 0; i < d; ++i)
         {
@@ -89,7 +89,7 @@ class BruteForce
 
     public:
 
-        BruteForce(py::array_t<double> points_py, std::string metric) : metric(metric)
+        BruteForce(py::array_t<float> points_py, std::string metric) : metric(metric)
         {
             py::buffer_info info = points_py.request();
             n = info.shape[0], d = info.shape[1];
@@ -98,29 +98,29 @@ class BruteForce
         std::string get_metric() const { return metric; }
 
         template <class DistanceFunctor>
-        std::tuple<std::vector<double>, std::vector<int64_t>, std::vector<int64_t>>
-        radius_neighbors_graph(const double *points, double radius, int num_threads) const
+        std::tuple<std::vector<float>, std::vector<int64_t>, std::vector<int64_t>>
+        radius_neighbors_graph(const float *points, float radius, int num_threads) const
         {
             static DistanceFunctor distance;
 
             omp_set_num_threads(num_threads);
 
             std::vector<int64_t> rowptrs(n+1), w(n,0), colids;
-            std::vector<double> dists;
+            std::vector<float> dists;
 
-            std::vector<std::tuple<int64_t, int64_t, double>> tuples;
+            std::vector<std::tuple<int64_t, int64_t, float>> tuples;
             int64_t *wptr = w.data();
 
             #pragma omp parallel shared(w, tuples)
             {
-                std::vector<std::tuple<int64_t, int64_t, double>> mytuples;
+                std::vector<std::tuple<int64_t, int64_t, float>> mytuples;
 
                 #pragma omp for nowait reduction(+:wptr[:n])
                 for (int64_t i = 0; i < n; ++i)
                 {
                     for (int64_t j = 0; j < n; ++j)
                     {
-                        double dist = distance(&points[j*d], &points[i*d], d);
+                        float dist = distance(&points[j*d], &points[i*d], d);
 
                         if (dist <= radius)
                         {
@@ -163,9 +163,9 @@ class BruteForce
 PYBIND11_MODULE(brute_force, m)
 {
     py::class_<BruteForce>(m, "brute_force")
-        .def(py::init<py::array_t<double>, std::string>())
+        .def(py::init<py::array_t<float>, std::string>())
         .def("radius_neighbors_graph",
-                [](const BruteForce& bf, py::array_t<double> points_py, double radius, int num_threads)
+                [](const BruteForce& bf, py::array_t<float> points_py, float radius, int num_threads)
                   {
                       std::string metric = bf.get_metric();
 
@@ -176,7 +176,7 @@ PYBIND11_MODULE(brute_force, m)
                       else throw std::runtime_error("Invalid metric!");
                   }
             );
-        /* .def("radius_neighbors_graph", [](const BruteForce& bf, py::array_t<double> points_py, double radius, int num_threads) { return bf.radius_neighbors_graph(npmem(points_py), radius, num_threads); }); */
+        /* .def("radius_neighbors_graph", [](const BruteForce& bf, py::array_t<float> points_py, float radius, int num_threads) { return bf.radius_neighbors_graph(npmem(points_py), radius, num_threads); }); */
 }
 
 /*
